requirements.txt
requests==2.22.0
pyquery==1.4.1
urllib3==1.25.7
Flask_Cors==3.0.8
ujson==1.35
websockets==8.1
cachetools==3.1.1
lxml==4.4.2
cfscrape==2.0.8
Flask==1.1.1
betfairlightweight==1.10.4
python-dotenv==0.10.3
websocket_client==0.56.0
pandas==1.2.3
